backend/grant-privilege/requirements.txt
psycopg2-binary==2.9.9
mcrcon==0.7.0
